+++ collective/auditlog/testing.py
@@ -9,9 +9,16 @@ from plone.app.testing import IntegrationTesting
 from zope.configuration import xmlconfig
 from plone.testing import z2
 
+try:
+    from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
+    BASE_FIXTURE = PLONE_APP_CONTENTTYPES_FIXTURE
+except ImportError:
+    from plone.app.testing import PLONE_FIXTURE
+    BASE_FIXTURE = PLONE_FIXTURE
+
 
 class AuditLog(PloneSandboxLayer):
-    defaultBases = (PLONE_FIXTURE,)
+    defaultBases = (BASE_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
         # load ZCML
@@ -24,10 +31,10 @@ class AuditLog(PloneSandboxLayer):
         # install into the Plone site
         applyProfile(portal, 'collective.auditlog:default')
         setRoles(portal, TEST_USER_ID, ('Member', 'Manager'))
-        workflowTool = getToolByName(portal, 'portal_workflow')
-        workflowTool.setDefaultChain('simple_publication_workflow')
-        workflowTool.setChainForPortalTypes(('File',),
-                                            'simple_publication_workflow')
+        # workflowTool = getToolByName(portal, 'portal_workflow')
+        # workflowTool.setDefaultChain('simple_publication_workflow')
+        # workflowTool.setChainForPortalTypes(('File',),
+        #                                     'simple_publication_workflow')
 
 
 AuditLog_FIXTURE = AuditLog()